Bruk Aspect og meterRegistry for å måle tid bruk
Er work in progress
